+++ src/test/java/domain/BowlingTest.java
@@ -1,5 +1,7 @@
 package domain;
 
+import domain.bowling.Bowling;
+import domain.player.Player;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
