+++ src/test/java/domain/NormalFrameTest.java
@@ -1,5 +1,6 @@
 package domain;
 
+import domain.frame.impl.NormalFrame;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
