+++ src/test/java/domain/PlayerTest.java
@@ -1,5 +1,6 @@
 package domain;
 
+import domain.player.Player;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.NullAndEmptySource;
